cleanup criteria package
Compile check that expressions in criteria package implement Expression
interface.

Move each expression and interface to its own dedicated file. This
should help to find things more easily. Working with interfaces is hard
and with files for each, we simplify things.
