+++ criteria/doc.go
@@ -1,3 +1,4 @@
-// Package criteria holds a representation of expression trees and code for their manipulation
-// This package serves to decouple the concrete query language from the execution of the queries against the database
+// Package criteria holds a representation of expression trees and code for
+// their manipulation This package serves to decouple the concrete query
+// language from the execution of the queries against the database
 package criteria
